backend/requirements.txt
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-SocketIO==4.3.0
greenlet==0.4.16
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.18.5
pkg-resources==0.0.0
pyserial==3.4
python-dotenv==0.13.0
python-engineio==3.13.0
python-socketio==4.6.0
six==1.15.0
Werkzeug==1.0.1
